+++ js/components/carousel.js
@@ -2,17 +2,15 @@ class Carousel {
   #directoryPath;
   #RIGHT = -1;
   #LEFT = 1;
-  constructor({numberOfImages, delayTime, path}) {
+  constructor({ numberOfImages, delayTime, carouselImagesPath }) {
     this.CAROUSEL_COUNT = numberOfImages;
     this.DELAY_TIME = delayTime;
-    this.#directoryPath = path;
+    this.#directoryPath = carouselImagesPath;
     this.slider = document.querySelector('.slider');
-    this.leftBtn = document.getElementById('carousel_left');
-    this.rightBtn = document.getElementById('carousel_right');
   }
 
   init() {
-    this.setImages()
+    this.setImages();
     this.slideAuto();
     this.clickSlide();
   }
@@ -20,16 +18,17 @@ class Carousel {
   setImages() {
     const imageFiles = Array.from({ length : this.CAROUSEL_COUNT }, (_, i) => i);
     const imageTemplate = imageFiles.reduce((template, number) => {
-      const img = `<img class="slide" src="${this.#directoryPath}/${number}.jpg" alt="carousel 배경이미지${number}">`;
-      template += img;
+      template += `<img class="slide" src="${this.#directoryPath}/${number}.jpg" alt="carousel 배경이미지${number}">`;
       return template;
     }, '');
     this.slider.insertAdjacentHTML('afterbegin', imageTemplate);
   }
 
   clickSlide() {
-    this.leftBtn.addEventListener('click', this.translateSlideHandler.bind(this));
-    this.rightBtn.addEventListener('click', this.translateSlideHandler.bind(this));
+    const $leftBtn = document.getElementById('carousel_left');
+    const $rightBtn = document.getElementById('carousel_right');
+    $leftBtn.addEventListener('click', this.translateSlideHandler.bind(this));
+    $rightBtn.addEventListener('click', this.translateSlideHandler.bind(this));
   }
 
   translateSlideHandler({ target }) {
